test: combine interlinks coverage in one rendered package
Use one synthetic package to exercise interlinks in docstrings and user-guide pages, including autolinking and external-source aliases. Reuse a single site build for the related rendered tests.
