+++ tests/test_gdg_rendered.py
@@ -7722,13 +7722,13 @@ def test_DED_numpy_seealso_desc_descriptions_render():
             )
 
 
-# ── gdtest_interlinks_prose tests ─────────────────────────────────────────────
+# ── gdtest_interlinks tests ────────────────────────────────────────────────
 
 
 @pytest.mark.dedicated
 def test_DED_interlinks_prose_pages_exist():
-    """gdtest_interlinks_prose: all exports have reference pages."""
-    pkg = "gdtest_interlinks_prose"
+    """gdtest_interlinks: reference pages exist for every export."""
+    pkg = "gdtest_interlinks"
     if not _has_rendered_site(pkg):
         pytest.skip(f"{pkg} not rendered")
 
@@ -7740,8 +7740,8 @@ def test_DED_interlinks_prose_pages_exist():
 @pytest.mark.dedicated
 @requires_bs4
 def test_DED_interlinks_prose_links_resolved():
-    """gdtest_interlinks_prose: interlinks in prose are resolved to <a> tags."""
-    pkg = "gdtest_interlinks_prose"
+    """gdtest_interlinks: docstring interlinks become <a> tags."""
+    pkg = "gdtest_interlinks"
     if not _has_rendered_site(pkg):
         pytest.skip(f"{pkg} not rendered")
 
@@ -7767,8 +7767,8 @@ def test_DED_interlinks_prose_links_resolved():
 @pytest.mark.dedicated
 @requires_bs4
 def test_DED_interlinks_prose_hrefs_valid():
-    """gdtest_interlinks_prose: interlink hrefs point to valid reference pages."""
-    pkg = "gdtest_interlinks_prose"
+    """gdtest_interlinks: interlink hrefs point to reference pages."""
+    pkg = "gdtest_interlinks"
     if not _has_rendered_site(pkg):
         pytest.skip(f"{pkg} not rendered")
 
@@ -7794,10 +7794,10 @@ def test_DED_interlinks_prose_hrefs_valid():
 
 @pytest.mark.dedicated
 def test_DED_interlinks_prose_publishes_its_inventory():
-    """gdtest_interlinks_prose: the site publishes objects.inv for other projects."""
+    """gdtest_interlinks: the site publishes objects.inv at its root."""
     from great_docs._sphinx_inventory import decode
 
-    pkg = "gdtest_interlinks_prose"
+    pkg = "gdtest_interlinks"
     if not _has_rendered_site(pkg):
         pytest.skip(f"{pkg} not rendered")
 
@@ -7808,6 +7808,142 @@ def test_DED_interlinks_prose_publishes_its_inventory():
     names = {e.name for e in inv.entries}
     assert f"{pkg}.DuckDBStore" in names, f"documented objects missing from {sorted(names)[:5]}"
     assert all(e.uri for e in inv.entries), "an entry has no uri to link to"
+
+
+def _ilu_skip():
+    pkg = "gdtest_interlinks"
+    if not _has_rendered_site(pkg):
+        pytest.skip(f"{pkg} not rendered")
+
+
+def _ilu_site():
+    return _site_dir("gdtest_interlinks")
+
+
+@pytest.mark.dedicated
+@requires_bs4
+def test_DED_interlinks_userguide_ug_pages_exist():
+    """gdtest_interlinks: user-guide pages exist."""
+    _ilu_skip()
+    ug = _ilu_site() / "user-guide"
+    assert (ug / "getting-started.html").exists(), "Getting Started page missing"
+    assert (ug / "advanced.html").exists(), "Advanced page missing"
+
+
+@pytest.mark.dedicated
+@requires_bs4
+def test_DED_interlinks_userguide_external_links_resolve():
+    """gdtest_interlinks: a configured source and its alias resolve."""
+    _ilu_skip()
+    soup = _load_html(_ilu_site() / "user-guide" / "external.html")
+
+    hrefs = [a.get("href", "") for a in soup.find_all("a", class_="gdls-link")]
+    external = [h for h in hrefs if h == "https://extdemo.example/docs/Widget.html"]
+    assert len(external) == 2, f"expected the full name and the alias to resolve, got {hrefs}"
+
+
+@pytest.mark.dedicated
+@requires_bs4
+def test_DED_interlinks_userguide_shortened_links():
+    """gdtest_interlinks: [](`~pkg.Name`) in user-guide resolves to links."""
+    _ilu_skip()
+    ug = _ilu_site() / "user-guide"
+    page = ug / "getting-started.html"
+    if not page.exists():
+        pytest.skip("getting-started.html not found")
+
+    soup = _load_html(page)
+    gdls_links = soup.find_all("a", class_="gdls-link")
+    link_texts = [a.get_text(strip=True) for a in gdls_links]
+
+    # Shortened references use short display names.
+    for expected in ("DuckDBStore", "BaseStore", "query()"):
+        assert expected in link_texts, (
+            f"Getting Started: shortened interlink {expected!r} not found. Found: {link_texts}"
+        )
+
+
+@pytest.mark.dedicated
+@requires_bs4
+def test_DED_interlinks_userguide_hrefs_relative():
+    """gdtest_interlinks: user-guide links use ../reference/ paths."""
+    _ilu_skip()
+    ug = _ilu_site() / "user-guide"
+    page = ug / "getting-started.html"
+    if not page.exists():
+        pytest.skip("getting-started.html not found")
+
+    soup = _load_html(page)
+    gdls_links = soup.find_all("a", class_="gdls-link")
+    for link in gdls_links:
+        href = link.get("href", "")
+        # From user-guide/*.html, href should start with ../reference/
+        assert href.startswith("../reference/"), (
+            f"User-guide interlink href {href!r} should be relative to ../reference/"
+        )
+
+
+@pytest.mark.dedicated
+@requires_bs4
+def test_DED_interlinks_userguide_full_qualified():
+    """gdtest_interlinks: [](`pkg.Name`) renders the fully qualified name."""
+    _ilu_skip()
+    ug = _ilu_site() / "user-guide"
+    page = ug / "advanced.html"
+    if not page.exists():
+        pytest.skip("advanced.html not found")
+
+    soup = _load_html(page)
+    gdls_links = soup.find_all("a", class_="gdls-link")
+    link_texts = [a.get_text(strip=True) for a in gdls_links]
+
+    assert "gdtest_interlinks.BaseStore" in link_texts, (
+        f"Advanced: full qualified 'gdtest_interlinks.BaseStore' not found. "
+        f"Found: {link_texts}"
+    )
+
+
+@pytest.mark.dedicated
+@requires_bs4
+def test_DED_interlinks_userguide_custom_text():
+    """gdtest_interlinks: custom link text remains unchanged."""
+    _ilu_skip()
+    ug = _ilu_site() / "user-guide"
+    page = ug / "advanced.html"
+    if not page.exists():
+        pytest.skip("advanced.html not found")
+
+    soup = _load_html(page)
+    gdls_links = soup.find_all("a", class_="gdls-link")
+    link_texts = [a.get_text(strip=True) for a in gdls_links]
+
+    assert "custom link text" in link_texts, (
+        f"Advanced: custom text 'custom link text' not found. Found: {link_texts}"
+    )
+    assert "custom text with tilde" in link_texts, (
+        f"Advanced: custom text 'custom text with tilde' not found. Found: {link_texts}"
+    )
+
+
+@pytest.mark.dedicated
+@requires_bs4
+def test_DED_interlinks_userguide_autolinked_code():
+    """gdtest_interlinks: inline code autolinks in user-guide pages."""
+    _ilu_skip()
+    ug = _ilu_site() / "user-guide"
+    page = ug / "advanced.html"
+    if not page.exists():
+        pytest.skip("advanced.html not found")
+
+    soup = _load_html(page)
+    # Autolinked code uses both the gdls-link and gdls-code classes.
+    code_links = soup.find_all("a", class_="gdls-code")
+    link_texts = [a.get_text(strip=True) for a in code_links]
+
+    for expected in ("BaseStore", "DuckDBStore", "query()"):
+        assert expected in link_texts, (
+            f"Advanced: autolinked code {expected!r} not found. Found: {link_texts}"
+        )
 
 
 # ═══════════════════════════════════════════════════════════════════════════════
@@ -9761,157 +9897,6 @@ def test_SCALE_MIN_SCALE_float_page_selectors():
     selectors = meta.get("data-selectors", "")
     for sel in ("#stf_wide", "#stf_styled", "#summary_card"):
         assert sel in selectors, f"Missing {sel} in float-override page selectors"
-
-
-# ═══════════════════════════════════════════════════════════════════════════════
-# DED: GDLS on non-reference pages (gdtest_interlinks_userguide)
-# ═══════════════════════════════════════════════════════════════════════════════
-
-
-def _ilu_skip():
-    pkg = "gdtest_interlinks_userguide"
-    if not _has_rendered_site(pkg):
-        pytest.skip(f"{pkg} not rendered")
-
-
-def _ilu_site():
-    return _site_dir("gdtest_interlinks_userguide")
-
-
-@pytest.mark.dedicated
-@requires_bs4
-def test_DED_interlinks_userguide_ref_pages_exist():
-    """gdtest_interlinks_userguide: all exports have reference pages."""
-    _ilu_skip()
-    ref = _ref_dir("gdtest_interlinks_userguide")
-    for name in ("Engine", "Connection", "execute"):
-        assert (ref / f"{name}.html").exists(), f"Ref page {name}.html missing"
-
-
-@pytest.mark.dedicated
-@requires_bs4
-def test_DED_interlinks_userguide_ug_pages_exist():
-    """gdtest_interlinks_userguide: user-guide pages exist."""
-    _ilu_skip()
-    ug = _ilu_site() / "user-guide"
-    assert (ug / "getting-started.html").exists(), "Getting Started page missing"
-    assert (ug / "advanced.html").exists(), "Advanced page missing"
-
-
-@pytest.mark.dedicated
-@requires_bs4
-def test_DED_interlinks_userguide_external_links_resolve():
-    """gdtest_interlinks_userguide: a configured source resolves, alias included."""
-    _ilu_skip()
-    soup = _load_html(_ilu_site() / "user-guide" / "external.html")
-
-    hrefs = [a.get("href", "") for a in soup.find_all("a", class_="gdls-link")]
-    external = [h for h in hrefs if h == "https://extdemo.example/docs/Widget.html"]
-    assert len(external) == 2, f"expected the full name and the alias to resolve, got {hrefs}"
-
-
-@pytest.mark.dedicated
-@requires_bs4
-def test_DED_interlinks_userguide_shortened_links():
-    """gdtest_interlinks_userguide: [](`~pkg.Name`) in user-guide resolves to links."""
-    _ilu_skip()
-    ug = _ilu_site() / "user-guide"
-    page = ug / "getting-started.html"
-    if not page.exists():
-        pytest.skip("getting-started.html not found")
-
-    soup = _load_html(page)
-    gdls_links = soup.find_all("a", class_="gdls-link")
-    link_texts = [a.get_text(strip=True) for a in gdls_links]
-
-    # Shortened references should produce short display names
-    for expected in ("Engine", "Connection", "execute()"):
-        assert expected in link_texts, (
-            f"Getting Started: shortened interlink {expected!r} not found. Found: {link_texts}"
-        )
-
-
-@pytest.mark.dedicated
-@requires_bs4
-def test_DED_interlinks_userguide_hrefs_relative():
-    """gdtest_interlinks_userguide: user-guide interlink hrefs use ../reference/ paths."""
-    _ilu_skip()
-    ug = _ilu_site() / "user-guide"
-    page = ug / "getting-started.html"
-    if not page.exists():
-        pytest.skip("getting-started.html not found")
-
-    soup = _load_html(page)
-    gdls_links = soup.find_all("a", class_="gdls-link")
-    for link in gdls_links:
-        href = link.get("href", "")
-        # From user-guide/*.html, href should start with ../reference/
-        assert href.startswith("../reference/"), (
-            f"User-guide interlink href {href!r} should be relative to ../reference/"
-        )
-
-
-@pytest.mark.dedicated
-@requires_bs4
-def test_DED_interlinks_userguide_full_qualified():
-    """gdtest_interlinks_userguide: [](`pkg.Name`) renders full qualified name."""
-    _ilu_skip()
-    ug = _ilu_site() / "user-guide"
-    page = ug / "advanced.html"
-    if not page.exists():
-        pytest.skip("advanced.html not found")
-
-    soup = _load_html(page)
-    gdls_links = soup.find_all("a", class_="gdls-link")
-    link_texts = [a.get_text(strip=True) for a in gdls_links]
-
-    assert "gdtest_interlinks_userguide.Engine" in link_texts, (
-        f"Advanced: full qualified 'gdtest_interlinks_userguide.Engine' not found. "
-        f"Found: {link_texts}"
-    )
-
-
-@pytest.mark.dedicated
-@requires_bs4
-def test_DED_interlinks_userguide_custom_text():
-    """gdtest_interlinks_userguide: [custom text](`pkg.Name`) preserves display text."""
-    _ilu_skip()
-    ug = _ilu_site() / "user-guide"
-    page = ug / "advanced.html"
-    if not page.exists():
-        pytest.skip("advanced.html not found")
-
-    soup = _load_html(page)
-    gdls_links = soup.find_all("a", class_="gdls-link")
-    link_texts = [a.get_text(strip=True) for a in gdls_links]
-
-    assert "custom link text" in link_texts, (
-        f"Advanced: custom text 'custom link text' not found. Found: {link_texts}"
-    )
-    assert "custom text with tilde" in link_texts, (
-        f"Advanced: custom text 'custom text with tilde' not found. Found: {link_texts}"
-    )
-
-
-@pytest.mark.dedicated
-@requires_bs4
-def test_DED_interlinks_userguide_autolinked_code():
-    """gdtest_interlinks_userguide: `Engine`, `execute()` autolinked in user-guide."""
-    _ilu_skip()
-    ug = _ilu_site() / "user-guide"
-    page = ug / "advanced.html"
-    if not page.exists():
-        pytest.skip("advanced.html not found")
-
-    soup = _load_html(page)
-    # Look for gdls-link gdls-code links (autolinked code produces this class combo)
-    code_links = soup.find_all("a", class_="gdls-code")
-    link_texts = [a.get_text(strip=True) for a in code_links]
-
-    for expected in ("Engine", "Connection", "execute()"):
-        assert expected in link_texts, (
-            f"Advanced: autolinked code {expected!r} not found. Found: {link_texts}"
-        )
 
 
 # ═══════════════════════════════════════════════════════════════════════════════
